third_party/WebKit/LayoutTests/fast/dom/shadow/tooltips-in-shadow-expected.txt

CONSOLE WARNING: line 87: Calling Element.createShadowRoot() for an element which already hosts a shadow root is deprecated and will be removed in M63, around December 2017. See https://www.chromestatus.com/features/4668884095336448 for more details.
Test for Case1: Non-Nested
PASS harness.expected is testRunner.tooltipText
Test for Case 2: Nested
PASS harness.expected is testRunner.tooltipText
Test for Case 3: Multiple Shadow
PASS harness.expected is testRunner.tooltipText
PASS successfullyParsed is true

TEST COMPLETE
